resolving review comments
